### 2018-02-20 #590 @whedon
[PRE REVIEW]: Phylen: automatic phylogenetic reconstruction using the EggNOG database
**Submitting author:** @giraola (<a href="http://orcid.org/0000-0002-6516-3404">Gregorio Iraola</a>)
**Repository:** <a href="https://github.com/iferres/phylen" target ="_blank">https://github.com/iferres/phylen</a>
**Version:** v0.1.0
