### 2018-02-20 7674ada
Update paper.md
