### 2018-02-20 ca9098f
Update paper.md
