+++ paper.md
@@ -28,6 +28,9 @@ Phylen has been already used by our group for building the _Helicobacter_ genus 
 
 ![A) Schematic workflow of Phylen. B) Phylogeny of _Epsilonproteobacteria_ obtained with the eggNOG database (eproNOG orthologs).](figure1.png)
 
+
+
+
 Phylen is useful to automatically build phylogenies at any desired taxonomic rank from any set of genes like taxon-specific markers or the whole core genome. Phylen depends on the R package phangorn [@phangorn] for phylogenetic reconstruction and external tools including HMMER3 [@hmmer3] as gene search engine and MAFFT [@mafft] for multiple sequence alignment.
 
 # Acknowledgements
