### 2018-02-20 05a0ff6
Update paper.md
