### 2018-02-21 a269313
changes in documentation and typos in paper.md
