### 2018-02-21 #590 @Kevin-Mattheus-Moerman
@KlausVigo would you or your co-workers be interested in reviewing this submission for JOSS (the journal of open source software)?

Review largely focuses on [the software project](https://github.com/iferres/phylen) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.00590/joss.00590/10.21105.joss.00590.pdf).

Background information:
- About [JOSS](http://joss.theoj.org/)
- [Reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines)
