### 2018-02-22 #590 @kortschak
I can do this, though R is not a preferred language for me so if one of the others want to pick this up that may be better. Having had a brief look through the code though, I don't see any obvious automated testing.
