### 2018-02-22 #593 @whedon
[REVIEW]: Phylen: automatic phylogenetic reconstruction using the EggNOG database
**Submitting author:** @giraola (<a href="http://orcid.org/0000-0002-6516-3404">Gregorio Iraola</a>)
**Repository:** <a href="https://github.com/iferres/phylen" target ="_blank">https://github.com/iferres/phylen</a>
**Version:** v0.1.0
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.6223538.v1" target="_blank">10.6084/m9.figshare.6223538.v1</a>

---
**REVIEWER 1**

---
**REVIEWER 2**

---
**REVIEWER 3**
