### 2018-02-22 #593 @Kevin-Mattheus-Moerman
@giraola @iferres can you comment on testing your code or point @kortschak to a "test suite"?
