### 2018-02-22 #593 @iferres
Thanks @Kevin-Mattheus-Moerman. @kortschak Thanks for your comment. I'm preparing a real case example to improve the README.md, I will add some data to the repo so it can be tested.
