### 2018-02-22 #593 @giraola
Hi @kortschak, @Kevin-Mattheus-Moerman and @juanvillada, thanks for accepting the revision of our manuscript. Following your recommendations we want to include a full test situation, however we think it has no sense to upload the full genomes into the repo (genomes are also public), so do you think pointing to a download link for the test data would be ok?
