### 2018-02-22 fe47a4a
fixed bug when returning path of downloaded file
