+++ README.md
@@ -1,89 +1,191 @@
-# phylen
-Automatic phylogenetic reconstruction using the EggNOG database (or others) hmm files.
+phylen
+================
 
-## Dependencies
-To work, it is required to have installed on your $PATH variable the following software:
- * [Mafft](http://mafft.cbrc.jp/alignment/software/)
- * [HMMER 3.1b2](http://hmmer.org/download.html)
-### Other external dependencies
+Standard workflow
+-----------------
 
- * A [EggNOG](http://eggnogdb.embl.de/#/app/downloads) HMM database file. It's not required to decompress (`untar`) those files before running the pipeline. Or alternatively...
+This tutorial begins with the extraction of toy data attached on this package. It consists on 10 genomes of the Campylobacterales order, 5 of them are Campylobacter species and the other 5 are Helicobacter species.
 
- * A custom set of HMMs concatenated into one single file.
- 
-## Installation
-```r
-devtools::install_github('iferres/phylen')
+``` r
+# List the attached tar.gz file
+tgz <- system.file('extdata', 'toydata.tar.gz', package = 'phylen')
+# List the files inside if it
+gff <- untar(tarfile = tgz, exdir = getwd(), list = T)
+# Decompress on current working directory
+untar(tarfile = tgz,exdir = getwd())
+# List the decompressed files
+gff
 ```
 
-## `phylen`: Compute Concatenated Alignment And Phylogeny
-### Usage
-```r
-phylen(gffs = character(), hmmFile = character(), isCompressed = TRUE,
-  eval = 1e-30, level, phyloMode = "ml", nbs = 100L, outDir = "phylen",
-  aliPfx = "coregenome", treePfx = "phylo", mafftMode = "linsi",
-  keepOgs = FALSE, n_threads = 1L)
+    ##  [1] "C_fetus_subsp_testudinum_Sp3.gff"                   
+    ##  [2] "C_fetus_subsp_venerealis_str_84-112.gff"            
+    ##  [3] "C_hyointestinalis_subsp_lawsonii_CCUG_27631.gff"    
+    ##  [4] "C_iguaniorum_str_RM11343.gff"                       
+    ##  [5] "C_pinnipediorum_subsp_pinnipediorum_str_RM17262.gff"
+    ##  [6] "H_bilis_str_AAQJH.gff"                              
+    ##  [7] "H_himalayensis_str_YS1.gff"                         
+    ##  [8] "H_pylori_J166.gff"                                  
+    ##  [9] "H_pylori_str_2018.gff"                              
+    ## [10] "H_typhlonius_str_1.gff"
+
+Lets load the `phylen` package and list the available Hidden Markov Models (HMMs) on the EggNOG database.
+
+``` r
+# Load the phylen package
+library(phylen)
 ```
-### Description
-Identify and align multiple genes, concatenate the alignments in a single file, and use it to compute a phylogeny.
 
-### Arguments
- * `gffs`: A `character` vector with the gff file paths.
- * `hmmFile`: The path to the ".hmm.tar.gz" file downloaded from EggNOG website or using `list_eggnogdb` and `download_nog_hmm` functions on this package. The already prepared ".hmm" text file can also be provided (see `isCompressed` below). Alternatively a custom set of ".hmm" files can be passed as a concatenated single file.
[…]
