### 2018-02-22 827cdd4
added short description to README.md
