### 2018-02-22 b10981f
typos and figure. README.md
