### 2018-02-22 #590 @KlausVigo
@Kevin-Mattheus-Moerman I could look at it but I have no time this week. Is there any potential conflict of interest as the package uses (and luckily cites) phangorn a lot?
