### 2018-02-22 #593 @iferres
@kortschak Please check the repo new README.md file. I have also attached example data, so you may have to re install the package.
