### 2018-02-22 #590 @Kevin-Mattheus-Moerman
@KlausVigo thanks. This review issue will remain open for a bit so we can wait for your to join after this week. Your input would be greatly appreciated!
Thanks for raising the issue of a conflicts of interest. I do not think them citing/using your work is an issue. In fact it further illustrates that you are a suitable reviewer as this directly applied to your expertise. We also have multiple reviewers for this submission. I'll add you as a reviewer as well and look forward to your comments. Let me know if you have any questions about the review process. It essentially involves ticking boxes at the top of this issue (I'll create a list for you).
