### 2018-02-22 #593 @Kevin-Mattheus-Moerman
@KlausVigo I've just added a tickbox set for you as well
