### 2018-02-22 #593 @Kevin-Mattheus-Moerman
@juanvillada (reviewer 1) @kortschak (reviewer 2) @KlausVigo (reviewer 3) we are good to go for the review process. Let me know if you are not able to tick the boxes at the top of this issue (be sure to accept this invite: https://github.com/openjournals/joss-reviews/invitations).
If you have questions just ping me or check out the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines). THANKS! :tada:
