### 2018-02-22 #593 @kortschak
@giraola Thank you for making a start on testing. I think that in the context of this tool integration testing is going to be very useful from a number of perspectives: 1) making people trust that the software behaves as advertised, 2) guards against regressions and 3) provides an idiomatic recipe for users to work from, so yes, I think that it should be OK to run a test on downloaded data (it might be worth letting the package do the download and cache it for convenience and efficiency), but I think that a smaller test set that is packaged with the code is also important. Extending this, while I can see that much of the code here is glue and doesn't lend itself well to unit testing, there are functions where that is easy to do, and adding unit tests will help you maintain to code in the future, so I'd suggest adding unit tests for most of the functions in `gff3_Handling.R` and some of the functions in `utils.R`.

Hadley Wickham has a nice package for testing that many people recommend: https://github.com/r-lib/testthat.
