### 2018-02-23 #593 @giraola
@kortschak thank you for these suggestions. We are exploring `testthat` to add unit tests to the main functions of `phylen`. In parallel, we already added a fully packaged example with step-by-step explanation to run it.
