### 2018-02-23 16852dc
minor changes to README
