### 2018-02-23 2838f46
starting testthat
