### 2018-03-02 #593 @kortschak
@giraola For the general checks, the question is "Has the submitting author (@giraola) made major contributions to the software?" Interpreting this literally it is hard to say yes based on `git blame`/`git log`, but I imagine that your input has been supervisional and I should be reading the author list as a biologist. Is this correct?

@Kevin-Mattheus-Moerman Can you clarify the interpretation of this question?
