### 2018-03-02 iferres/phylen#7 @kortschak
Please add list of dependencies needed for installation
In assessing whether the package satisfies the installability criterion for https://github.com/openjournals/joss-reviews/issues/593 I ran into a small road bump (I don't use phangorn, which is why this arose).

The install process halted with an error in installing igraph because of a lack of libxml2-dev on my system. It would be helpful for users if you provided a list of required system packages for each of the supported platforms since the R build systems are pretty unhelpful in pointing users in the right direction when things like this happen.
