### 2018-03-02 iferres/phylen#8 @kortschak
Please add community guidelines documentation
For https://github.com/openjournals/joss-reviews/issues/593, can you please add something that satisfies "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

A CONTRIBUTING.md and text in the README.md would be the best IMO.
