### 2018-03-02 iferres/phylen#9 @kortschak
README: figure 1A shows 4 nodes in a 3 taxon alignment
This is for https://github.com/openjournals/joss-reviews/issues/593, and is a bit of phylogenetic pedantry on my part.

Can you change the figure 1A to either show a simple 3 taxon star tree or, preferably, add a fourth taxon to the figure to make the tree interesting.
