### 2018-03-02 iferres/phylen#11 @kortschak
paper.md,README: need a statement of need
This is for https://github.com/openjournals/joss-reviews/issues/593.

A review criterion is a statement of need: "Do the authors clearly state what problems the software is designed to solve and who the target audience is?"

Can you please make a clearer statement to this effect in paper.md and README.md.
