### 2018-03-02 iferres/phylen#12 @kortschak
paper.bib: add DOIs for papers
This is for https://github.com/openjournals/joss-reviews/issues/593.

Please add DOIs for papers where they exist.
