### 2018-03-04 #593 @Kevin-Mattheus-Moerman
Hi @kortschak thanks for your question. In this case there are two contributors both of which are authors and appear to agree on their joint authorship. The check mark you mention really serves to weed out submission where somebody has for instance forked somebody else's work and submitted as main author without having made any significant contributions. Other contributors might object to such a person claiming main authorship. Anyway, this is clearly not the case here.
