### 2018-03-04 #593 @kortschak
Thanks @Kevin-Mattheus-Moerman. Maybe the question could be reworded to something like "Has the authors made major contributions to the software?"
