### 2018-03-04 #593 @giraola
Hi @kortschak thanks for your question. Indeed, @iferres mainly contributed to develop the code within the repository, but both have made major contributions to the software. I am @iferres's Master thesis supervisor and my contributions were mainly conceptual.
