### 2018-03-06 iferres/phylen#12 @giraola
DOIs have been added to paper.bib entries.
