### 2018-03-06 iferres/phylen#11 @giraola
We have expanded the last paragraph (before Acknowledgements section) to better explain what problems the software is designed to solve and who is the target audience.
