### 2018-03-06 iferres/phylen#9 @giraola
We have updated the figure.
