### 2018-03-06 iferres/phylen#9 @kortschak
It looks like the png file was deleted in the repo rather than modified, leaving the link in the paper broken.
