### 2018-03-06 4f0488d
Update paper.md
