+++ paper.md
@@ -33,6 +33,6 @@ Phylen has been designed to facilitate the reconstruction of high-resolution phy
 
 # Acknowledgements
 
-I.F. was supported by ANII (Uruguay) posgraduation grant POS_NAC_2016_1_131079. We thank Pablo Fresia and Daniela Costa for testing Phylen. 
+I.F. was supported by ANII (Uruguay) postgraduation grant POS_NAC_2016_1_131079. We thank Pablo Fresia and Daniela Costa for testing Phylen. 
 
 # References
