### 2018-03-07 e4cfe2b
Create CONTRIBUTING.md
