+++ CONTRIBUTING.md
@@ -1,0 +1,11 @@
+# Code of Conduct
+
+Maintainers and other contributors must follow this repository's [code of conduct](https://github.com/iferres/phylen/blob/master/CODE_OF_CONDUCT.md).
+
+# Suggestions and bug reports
+
+`phylen` thrives on the suggestions and bug reports you submit to the [issue tracker](https://github.com/iferres/phylen/issues). Before posting, please search both the open and closed issues to help us avoid duplication.
+
+# Patches to code and documentation
+
+Please help us develop `phylen`. To submit a patch, please [fork this repository](https://help.github.com/articles/fork-a-repo/) and submit a [pull request](https://github.com/iferres/phylen/pulls).
