### 2018-03-07 iferres/phylen#8 @iferres
Added CODE_OF_CONDUCT.md, CONTRIBUTING.md and some text at README.md . Do you think it's ok? @kortschak
