### 2018-03-07 iferres/phylen#9 @kortschak
Strictly speaking it is still not a possible inference from the hypothetical input data as the rooting requires an out group, but its really just pedantry on my part, so I'll leave this since it's really only intended to be illustrative.
