### 2018-03-08 4bf56c5
added reference to prokka on input files
