### 2018-03-08 fffa729
added reference to prokka on input files
