### 2018-03-16 #593 @kortschak
@giraola I will be able to look at the functionality criterion when the tests are added, so after that and the addition of the correct version tag, I'll be able to complete my review.
