### 2018-03-19 acc937e
deleted unused function
