### 2018-03-20 1938ef7
imporved test for gff handling
