+++ tests/testthat/test_gff3.R
@@ -1,44 +1,57 @@
 context('gff3')
 
-source(system.file('R', 'gff3_Handling.R', package = 'phylen'))
+tgz <- system.file('testdata', 'test.tar.gz', package = 'phylen')
 
-tgz <- system.file('extdata', 'toydata.tar.gz', package = 'phylen')
-
-file <- 'H_bilis_str_AAQJH.gff'
-untar(tgz, files = , exdir = tempdir())
-gff <- list.files(path = tempdir(),
-                  pattern = file,
+untar(tgz, files = 'test.gff', exdir = tempdir())
+tesf_gff <- list.files(path = tempdir(),
+                  pattern = '^test_gff$',
                   full.names = TRUE)
 
-rl <- readLines(gff)
-fasta.ini <- grep("##FASTA",rl) + 1
-fasta.end <- length(rl)
-fasta <- rl[fasta.ini:fasta.end]
-fheaders <- grep('>',fasta,fixed = T)
-t <- matrix(nrow = length(fheaders),ncol = 2)
-t[,1] <- fheaders + 1
-t[,2] <- c((fheaders - 1)[-1],length(fasta))
-ap <- apply(t,1,function(x){
-  paste(fasta[x[1]:x[2]],collapse = '')
-})
-names(ap) <- gsub('>','',fasta[fheaders])
-fnas <- lapply(tolower(ap),function(x){
-  seqinr::as.SeqFastadna(seqinr::s2c(x),name = names(x))
-})
+gff_table_df_file <- system.file('testdata',
+                                 'gff_table_df.rds',
+                                 package = 'phylen')
+gff_table_df <- readRDS(gff_table_df_file)
+
 
 test_that("extracting table from gff works",{
+  rl <- readLines(gff)
   tt <- phylen:::extractGffTable(rl)
   d <- dim(tt)
   d1 <- d[1]
   d2 <- d[2]
   expect_is(tt, class = "data.frame")
-  expect_equal(d1, 2365)
+  expect_equal(d1, 236)
   expect_equal(d2, 10)
+  expect_identical(tt, gff_table_df)
 })
 
+
+gene_SeqFastadna_file <- system.file('testdata',
+                                     'gene_SeqFastadna.rds',
+                                     package = 'phylen')
+gene_SeqFastadna <- readRDS(gene_SeqFastadna_file)
+
+gene_SeqFastaAA_file <- system.file('testdata',
+                                     'gene_SeqFastaAA.rds',
+                                     package = 'phylen')
+gene_SeqFastaAA <- readRDS(gene_SeqFastaAA_file)
+
+
+test_that("translate works", {
+  tr <- phylen:::translate(gene_SeqFastadna, numcode = 11)
+  expect_equivalent(tr, unclass(gene_SeqFastaAA))
+})
+
+
+SeqFastadna_file <- system.file('testdata',
+                                'SeqFastadna.rds',
+                                package = 'phylen')
+SeqFastadna <- readRDS(SeqFastadna_file)
+
+
 test_that("extracting sequences from gff table works",{
-  cds <- gfftable[1, ]
-  sq <- phylen:::getFfnFaa(fnas,
+  cds <- gff_table_df[1, ]
+  sq <- phylen:::getFfnFaa(SeqFastadna,
                            contig = cds$Contig,
                            strand = cds$Strand,
                            from = cds$From,
@@ -50,7 +63,9 @@ test_that("extracting sequences from gff table works",{
   expect_length(sq, 2)
   expect_is(sq[[1]], "SeqFastadna")
   expect_is(sq[[2]], "SeqFastaAA")
[…]
