+++ man/getSeqOfType.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gff3_Handling.R
-\name{getSeqOfType}
-\alias{getSeqOfType}
-\title{Get sequences (nucleotidic or aminoacidic)}
-\usage{
-getSeqOfType(seqs, type = "AA")
-}
-\arguments{
-\item{seqs}{\code{list} of sequences as passed by \link{extractSeqsFromGff3}.}
-
-\item{type}{\code{character}. Either 'AA' or 'DNA'.}
-}
-\value{
-A \code{list} of sequences.
-}
-\description{
-Extract cds either nucleotidic (gene) or aminoacidic (protein).
-}
-\author{
-Ignacio Ferres
-}
