### 2018-03-20 b515a9b
fixed bug on download_nog_hmm
