### 2018-03-20 c432a14
added getIdsFromStats test
