### 2018-03-21 83bec1c
added readDomtblout test
