### 2018-03-21 #593 @giraola
@kortschak @juanvillada tests are already added, you can have a look now in the repo. @kortschak which is wrong with the version tag?
