### 2018-03-21 #593 @juanvillada
Thanks @giraola, already did a preliminary "manual" test. Will do the proper test asap and post a couple of comments later this week.
