### 2018-03-21 #593 @iferres
@juanvillada There was a typo on one of the test scripts which was making the test to fail, now everything is working.
