### 2018-03-22 #593 @kortschak
I am unable to run the tests successfully:
```
> library(testthat)
> library(phylen)
Loading required package: phangorn
Loading required package: ape
> test_package(phylen)
Error in match(x, table, nomatch = 0L) :
  'match' requires vector arguments
> sessionInfo()
R version 3.4.4 (2018-03-15)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 16.04.4 LTS

Matrix products: default
BLAS: /usr/lib/atlas-base/atlas/libblas.so.3.0
LAPACK: /usr/lib/atlas-base/atlas/liblapack.so.3.0

locale:
 [1] LC_CTYPE=en_AU.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_AU.UTF-8        LC_COLLATE=en_AU.UTF-8
 [5] LC_MONETARY=en_AU.UTF-8    LC_MESSAGES=en_AU.UTF-8
 [7] LC_PAPER=en_AU.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_AU.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] phylen_0.1.0   phangorn_2.4.0 ape_5.0        testthat_2.0.0

 [1] Rcpp_0.12.16    quadprog_1.5-5  lattice_0.20-35 XML_3.98-1.10
 [5] MASS_7.3-40     grid_3.4.4      R6_2.2.2        nlme_3.1-131.1
 [9] magrittr_1.5    httr_1.3.1      rlang_0.2.0     seqinr_3.4-5
[13] Matrix_1.2-11   fastmatch_1.1-0 ade4_1.7-10     igraph_1.2.1
[17] parallel_3.4.4  compiler_3.4.4  pkgconfig_2.0.1
```

It might be worthwhile setting up a CI service job with TravisCI or similar so that tests are run on commit.
