### 2018-03-22 #593 @iferres
I think "phylen" should be between quotes in test_package function. Anyway, I still got an error:
``
> test_package("phylen")
Error: No tests found for phylen

``
which is strange because it runs ok when I use Ctrl + Shift + T inside RStudio. Let me check what's going on and I let you know.
I will check the CI stuff.
Thanks for all the recommendations, by the way, its my first package on revision and I'm learning a lot with your suggestions.
