### 2018-03-22 #593 @iferres
Regarding version: Is written at the DESCRIPTION file. Do you think it should be more visible?
