### 2018-03-22 #593 @kortschak
@iferres Thank you for doing that. Testthat is not wildly helpful in this kind of context I see.

For the version, at some point you will need to make a release, github will require that this is a tagged commit and the JOSS requirement is that the release matches version here "Does the release version given match the GitHub release (v0.1.0)?" This is obviously impossible if followin good git tag hygiene and changes are requested, but after everyone is happy with the state of the code/text, then you should tage the HEAD with v0.1.0 and make that the release for your doi store submission.

Thank you for bearing with me here.
