### 2018-03-22 #593 @iferres
@kortschak Ok, done. Thanks for your patience explaining this. Please, confirm if what I did is what is needed.
Regarding the `test_package` issue, I found [this](https://github.com/r-lib/testthat/issues/205#issuecomment-69342965).
