### 2018-03-22 #593 @kortschak
That works (hopefully there won't need to be a bump, but it looks like there won't be - in my submissions it has gone in as v1.0.0 and become v1.0.1 by the time review has gone through).

Thanks for the link (I filed an issue at testthat asking for improvement of the docs).
