### 2018-03-23 iferres/phylen#7 @iferres
Hi @kortschak . Sorry for the delay answering this thread. I found other people have had the same problem with the igraph package installation. I added a paragraph at README.md pointing out the igraph system requirements. Thanks for the note.

By the way, could you finally install and test the functionality of phylen?
