### 2018-03-23 iferres/phylen#7 @kortschak
Thank you, that addresses it.

Yes, I' managed to install the software, but I was waiting for unit tests before I go ahead an look at functionality. I'll hopefully get to that over the next week.
