### 2018-04-05 31b0b03
use NNI rearrangement as default with ML
