### 2018-04-05 #593 @KlausVigo
Hi @giraola and @iferres,

I just made a pull request, to make the default closer to what users would expect. The `optim.pml` should also some tree rearrangements, otherwise it would just return an optimised NJ tree. Additionally I added some code which sets the number of rate categories to 4 if optimisation of the discrete gamma distribution is requested. That is probably some of the more quirky things of using `pml` and `optim.pml` in phangorn. You want to test that it works properly.

I like that you added travis-ci support. It is actually useful to try also eliminate all notes and warnings messages from R CMD check. The other small commit was just doing that. You may have a copy of the paper.md also on the vignettes folder and the file `readme_img1.png` in that folder seems to be just a relict of an older document.

This is not really the objective of this review, but I think it should be possible to distribute the package to other operation systems not just linux with only minor changes. You would only need an option to set the path to the binaries or ensure that the binaries are on the search path from hmmr, mafft. It is done in ape for the example in the muscle function. It still is a nightmare if you give a course and half the students don't know where their binaries are sitting on the disc. This would allow to upload the package on CRAN and to reach a wider audience of users.

Regards,
Klaus
