### 2018-04-05 #593 @giraola
Hi @KlausVigo thanks for your time and for optimising the tree building step. We will have a look at muscle() function in APE. We agree that would be better to distribute the package to other OSs, indeed some colleagues have already requested this!
