### 2018-04-06 #593 @juanvillada
Hi @giraola  and @iferres,

### Summary of my review:
- ✅ General Checks
- ⚠️ Functionality
- ✅ Documentation
- ✅ Software paper
----

## ✅ General Checks
I had two comments for this section on **Version** and **Authorship**. They were addressed already.

## ⚠️ Functionality
### Installation
- Assuming some people do not have it previously installed, I would suggest to include a line to install the `devtools` package.

```r
install.packages("devtools")
devtools::install_github("iferres/phylen")
```

**Minor**: I don't think you must include this in the installation docs, but only in case you find this useful, I got `phylen` to work on a new and clean Ubuntu distro only after doing this:
```r
install.packages("git2r")
install.packages("httr")
install.packages("devtools")
install.packages("XML")
devtools::install_github("iferres/phylen")
library(phylen)
```

- In order to avoid people getting stuck with requirements, I would suggest to include some lines in order to faciliate `MAFFT` and `HMMER` installation:

Assuming homebrew (macOS) or linuxbrew (Linux) availability:

```sh
brew install brewsci/science/mafft
brew install brewsci/science/hmmer
```
or with `apt`:

```sh
sudo apt install mafft
sudo apt install hmmer
```

- I tested `phylen` with `testthat`. It is OK.
```r
library(phylen)
install.packages("testthat")
library(testthat)
test_package("phylen")
> test_package("phylen")
══ testthat results  ════════════════════════════════════════════════════════════════════════════
OK: 23 SKIPPED: 0 FAILED: 0
```

- ❗️**Manual test**. I have not been able to contruct the final tree following the instructions in `README.md`, see the output below:

<details><summary>RStudio console output</summary>

```r
> p <- phylen(gffs = gff, # The gff files extracted on the first step
+             hmmFile = hmm, # The downloaded HMMs
+             isCompressed = TRUE, # The downloaded HMMs are compressed (tar.gz)
+             phyloMode = "ml", # nj or ml, in this case we will use maximum likelihood
+             nbs = 100, # The number of bootstrap.
+             level = 100, # The percentage of genomes a gene must be in to be considered as part of the coregenome.
+             outDir = "phylen", # Lets create a directory called "phylen" to put the output files
+             n_threads = 2) # Use 1 threads
Decompressing.. DONE!
Concatenating..
===============================================================================================================
DONE!
Pressing.. DONE!
Getting information from hmms.. DONE!
Searching.. DONE!
Computing panmatrix.. DONE!
Getting core-genes.. DONE!
Writting fastas.. DONE!
Aligning.. sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not foundsh: 1:
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: sh: 1: linsi: not found
linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
sh: 1: linsi: not found
DONE!
Concatenating.. DONE!
Removing intermediate files.. DONE!
Generating phylogeny..
Error in data[[1]] : subscript out of bounds
In addition: There were 22 warnings (use warnings() to see them)
> p
Error: object 'p' not found
```

## ✅ Documentation
- It would be nice to add some [**topics**](https://help.github.com/articles/classifying-your-repository-with-topics/) "_to categorize your repository and make it more discoverable_”.

## ✅ Software paper
Seems fine now after DOIs correction  👍
