### 2018-04-06 476fd09
documentation of patched compute phylo with rearrangement param
