### 2018-04-06 cbc7fbf
Added reference to mafft aliases installation at readme and manual pages
