### 2018-04-06 f7ee116
minor verbose fix
