### 2018-04-06 7e722d3
added more options for controling phylogeny inference
