### 2018-04-06 #593 @KlausVigo
Hi @iferres ,

I think these changes would be enough for the start. I didn't want to suggest that you need to add the sources of HMMER and MAFFT and install them with the package, that is probably too much work. If you could be done it is super useful of course.

There is also the `msa` package on Bioconductor, which installs muscle and clustal to use from inside R. If the maintainer would add MAFFT, this would be really great and one would have all alignment programs in one package. The maintainer Ulrich Bodenhofer seems a really nice guy (and he got Ripleyed because of me), so may try to get in contact with him. And there are some function in `phangorn` / `ape` to convert between the different R alignment formats.

Have a nice weekend

Cheers,
Klaus
