### 2018-04-20 #593 @giraola
Hi @Kevin-Mattheus-Moerman, we would like to know what additional specific modifications are needed to fulfil reviewer's requirements. These are boxes that remain unchecked:

- @juanvillada > Functionality
- @kortschak > Functionality
- @KlausVigo > Version, Statement of need, Functionality documentation.

Cheers,
Gregorio
