### 2018-04-21 #593 @juanvillada
Hi @giraola and @iferres!

I reinstalled and tried `phylen` again today. It is working as expected.
Functionality OK!

```r
> devtools::install_github("iferres/phylen")
> library(phylen)
> tgz <- system.file('extdata', 'toydata.tar.gz', package = 'phylen')
> gff <- untar(tarfile = tgz, exdir = getwd(), list = T)
> untar(tarfile = tgz,exdir = getwd())
> hmm <- download_nog_hmm(nog.prefix = "eproNOG", onDir = getwd())
> p <- phylen(gffs = gff, # The gff files extracted on the first step
            hmmFile = hmm, # The downloaded HMMs
            isCompressed = TRUE, # The downloaded HMMs are compressed (tar.gz)
            phyloMode = "ml", # nj or ml, in this case we will use maximum likelihood
            nbs = 100, # The number of bootstrap.
            level = 100, # The percentage of genomes a gene must be in to be considered as part of the coregenome.
            outDir = "phylen", # Lets create a directory called "phylen" to put the output files
            n_threads = 2) # Use 2 threads

Decompressing.. DONE!
Concatenating..
==============================================================================================================
DONE!
Pressing.. DONE!
Getting information from hmms.. DONE!
Searching.. DONE!
Computing panmatrix.. DONE!
Getting core-genes.. DONE!
Writting fastas.. DONE!
Aligning.. DONE!
Concatenating.. DONE!
Removing intermediate files.. DONE!
Generating phylogeny..
optimize edge weights:  -5958264 --> -5731301
optimize base frequencies:  -5731301 --> -5685751
optimize rate matrix:  -5685751 --> -5620789
optimize edge weights:  -5620789 --> -5619310
optimize topology:  -5619310 --> -5619310
0
optimize base frequencies:  -5619310 --> -5615981
optimize rate matrix:  -5615981 --> -5614940
optimize edge weights:  -5614940 --> -5614927
optimize base frequencies:  -5614927 --> -5614698
optimize rate matrix:  -5614698 --> -5614643
optimize edge weights:  -5614643 --> -5614640
optimize base frequencies:  -5614640 --> -5614626
optimize rate matrix:  -5614626 --> -5614623
optimize edge weights:  -5614623 --> -5614623
optimize base frequencies:  -5614623 --> -5614621
optimize rate matrix:  -5614621 --> -5614621
optimize edge weights:  -5614621 --> -5614621
optimize base frequencies:  -5614621 --> -5614621
optimize rate matrix:  -5614621 --> -5614621
optimize edge weights:  -5614621 --> -5614621
Generating phylogeny.. DONE!

Finished: 658 groups of orthologous from 10 isolates have been used in the alignment.
Returning an object of class "phylo" with 10tips and 8 nodes.
> p

Phylogenetic tree with 10 tips and 8 internal nodes.

Tip labels:
	C_fetus_subsp_testudinum_Sp3, C_fetus_subsp_venerealis_str_84-112, C_hyointestinalis_subsp_lawsonii_CCUG_27631, C_iguaniorum_str_RM11343, C_pinnipediorum_subsp_pinnipediorum_str_RM17262, H_bilis_str_AAQJH, ...
Node labels:
	100, 100, 100, 100, 100, 100, ...

Unrooted; includes branch lengths.
> class(p)
[1] "phylo"
> plot(p, type = 'unrooted', cex = 0.7, lab4ut = 'axial')
```
