### 2018-05-01 #593 @kortschak
Sorry for the delay. All accepted now.
