### 2018-05-02 #593 @giraola
Thanks @kortschak, @juanvillada and @KlausVigo for this constructive review process. @Kevin-Mattheus-Moerman, it seems every box has been checked by the three reviewers. What's next?
