### 2018-05-03 #593 @Kevin-Mattheus-Moerman
@giraola @iferres what is next is that I will congratulate you as I will now recommend that this work is accepted! :tada: :rocket: :robot:
At this point, can you please make an archive of the final reviewed software in [Zenodo](https://zenodo.org/), or [figshare](https://figshare.com/), or another service, and update this thread with the DOI of the archive? Once we have that available we can process acceptance.
