### 2018-05-04 #593 @Kevin-Mattheus-Moerman
Arfon, over to you, we are all set. The authors say the DOI is 10.6084/m9.figshare.6223538. Although the link has .v1 in it see here: [https://doi.org/10.6084/m9.figshare.6223538.v1](https://doi.org/10.6084/m9.figshare.6223538.v1), so make sure we'll use the correct one.
